# Avoid unnecessary copies during config generation